# custos/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  internal_interface_http_handler.CreateTenantRequest:
    properties:
      domain:
        type: string
      max_users:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
    required:
    - name
    - slug
    type: object
...
  internal_interface_http_handler.TenantResponse:
    properties:
      created_at:
        type: string
      domain:
        type: string
      id:
        type: integer
      max_users:
        type: integer
      name:
        type: string
      plan:
        type: string
      slug:
        type: string
      status:
        type: string
      updated_at:
        type: string
      user_count:
        type: integer
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@fly.com
    name: API Support
    url: http://www.fly.com/support
  description: 用户认证和授权服务API文档，提供用户注册、登录、OAuth、RBAC等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Custos User & Authentication Service API
  version: "1.0"
paths:
...
  /tenants:
    get:
      description: 获取所有租户列表，支持分页（需要管理员权限）
      parameters:
      - default: 20
        description: 每页数量（1-100）
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              limit:
                type: integer
              offset:
                type: integer
              tenants:
                items:
                  $ref: '#/definitions/internal_interface_http_handler.TenantResponse'
                type: array
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 租户列表
      tags:
      - 租户
    post:
      consumes:
      - application/json
      description: 创建新的租户（需要管理员权限）
      parameters:
      - description: 租户信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_interface_http_handler.CreateTenantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_interface_http_handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 创建租户
      tags:
      - 租户
  /tenants/{id}:
    get:
      description: 根据ID获取租户详细信息
      parameters:
      - description: 租户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interface_http_handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取租户信息
      tags:
      - 租户
  /tenants/current:
    get:
      description: 获取当前上下文中的租户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_interface_http_handler.TenantResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取当前租户
      tags:
      - 租户
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...